feat: add delay for backend propagation in MKCOL handler and update tests
